Integration test node follows to phase 0
